% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NGSproj.data-package.R
\docType{package}
\name{NGSproj.data-package}
\alias{NGSproj.data-package}
\title{Xiaobei's collected datasets for NGS projects}
\description{
Xiaobei's collected datasets for NGS projects
}
\details{
Listed are a few highlighted sets
\itemize{
\item \code{\link{...}}
}

You may find the latest version of \code{NGSproj.data} at \url{https://github.com/xiaobeizhao/NGSproj.data}.
}
\references{
Xiaobei Zhao (2022)
}
\author{
Xiaobei Zhao
}
\keyword{package}
