% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test}
\alias{test}
\title{test dataset}
\format{
A data table with x rows and y columns:
\itemize{
  \item{}
  \item{}
  \item{}
}
}
\source{
\url{...}
}
\usage{
data(test)
}
\value{
data.table
}
\description{
test dataset
}
\details{
A dataset containing ...
}
\author{

}
\keyword{datasets}
